+++ src/main/java/com/trackery/trackerybackapiserver/domain/location/dto/MapResponseDto.java
@@ -1,6 +1,6 @@
-package com.trackery.trackerybackapiserver.domain.map.dto;
+package com.trackery.trackerybackapiserver.domain.location.dto;
 
-import com.trackery.trackerybackapiserver.domain.map.entity.Sigungu;
+import com.trackery.trackerybackapiserver.domain.location.entity.JusoSigungu;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -8,7 +8,7 @@ import lombok.Data;
 import lombok.NoArgsConstructor;
 
 /**
- * packageName    : com.trackery.trackerybackapiserver.domain.map.dto
+ * packageName    : com.trackery.trackerybackapiserver.domain.location.dto
  * fileName       : MapResponseDto
  * author         : inari
  * date           : 25. 5. 30.
@@ -46,13 +46,13 @@ public class MapResponseDto {
 	private Long sigunguId;
 
 	/**
-	 * Sigungu 엔티티로부터 DTO를 생성하는 정적 팩토리 메서드입니다.
+	 * JusoSigungu 엔티티로부터 DTO를 생성하는 정적 팩토리 메서드입니다.
 	 *
 	 * @param sigungu 시군구 엔티티 (시도 정보 포함)
 	 * @return MapResponseDto 인스턴스
-	 * @apiNote Sigungu 엔티티는 연관된 Sido 정보를 포함해야 합니다.
+	 * @apiNote JusoSigungu 엔티티는 연관된 JusoSido 정보를 포함해야 합니다.
 	 */
-	public static MapResponseDto from(Sigungu sigungu) {
+	public static MapResponseDto from(JusoSigungu sigungu) {
 		return MapResponseDto.builder()
 			.sidoName(sigungu.getSido().getSidoName())
 			.sigunguName(sigungu.getSigunguName())
