+++ src/main/java/com/trackery/trackerybackapiserver/domain/location/mapper/LocationMapper.java
@@ -1,12 +1,15 @@
 package com.trackery.trackerybackapiserver.domain.location.mapper;
 
-import java.util.Optional;
-
 import org.apache.ibatis.annotations.Mapper;
 import org.locationtech.jts.geom.Point;
 
+import java.util.List;
+import java.util.Optional;
+
 import com.trackery.trackerybackapiserver.domain.location.dto.CoordinateDto;
+import com.trackery.trackerybackapiserver.domain.location.dto.CoordinateRequestDto;
 import com.trackery.trackerybackapiserver.domain.location.entity.CoordinatePoint;
+import com.trackery.trackerybackapiserver.domain.location.entity.JusoSido;
 import com.trackery.trackerybackapiserver.domain.location.entity.JusoSigungu;
 
 /**
@@ -41,4 +44,29 @@ public interface LocationMapper{
 	 * @param coordinatePoint CoordinatePoint 객체
 	 */
 	void insertCoordinatePoint(CoordinatePoint coordinatePoint);
+
+	/**
+	 * 대한민국의 모든 시도 목록을 조회합니다.
+	 */
+	List<JusoSido> findAllSido();
+
+	/**
+	 * 특정 시도에 속한 모든 시군구 목록을 조회합니다.
+	 */
+	List<JusoSigungu> findSigunguBySido(Long sidoId);
+
+	/**
+	 * 지리 좌표를 기반으로 해당 위치의 시군구 정보를 조회합니다.
+	 */
+	JusoSigungu findSigunguByCoordinateRequest(CoordinateRequestDto coordinate);
+
+	/**
+	 * 특정 시도의 경계선을 GeoJSON 형식으로 조회합니다.
+	 */
+	String findSidoBorderAsGeoJson(Long sidoId);
+
+	/**
+	 * 특정 시군구의 경계선을 GeoJSON 형식으로 조회합니다.
+	 */
+	String findSigunguBorderAsGeoJson(Long sigunguId);
 }
