--- src/main/java/com/trackery/trackerybackapiserver/domain/map/entity/Sido.java
@@ -1,24 +1,0 @@
-package com.trackery.trackerybackapiserver.domain.map.entity;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-/**
- * packageName    : com.trackery.trackerybackapiserver.domain.map.entity
- * fileName       : Sido
- * author         : inari
- * date           : 25. 5. 20.
- * description    :
- * ===========================================================
- * DATE              AUTHOR             NOTE
- * -----------------------------------------------------------
- * 25. 5. 20.        inari       최초 생성
- */
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-public class Sido {
-	private Long sidoId;
-	private String sidoName;
-}
